HS-1007 Create help configuring own component
